pyproject.toml
[project]
name = "lg-agentic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.29.5",
    "langchain[anthropic,azure,community]>=0.3.25",
    "langgraph>=0.4.3",
    "python-dotenv>=1.1.0",
]
